Add asynchronous solver support
Adds support to run the maze solver in a separated thread. This is done
by calling maze_solve_thread() that is responsible for creating and
starting the solver thread. A callback can be specified so the caller
gets notified when the maze is solved. The callback is also called
periodically to allow for UI updates. The callback takes a 'reason'
parameter indicating the reason of the call:
 - SOLVER_CB_REASON_RUNNING: The solver is still running. This allows
   the caller to do some GUI updates.
 - SOLVER_CB_REASON_SOLVED: The maze is solved
 - SOLVER_CB_REASON_CANCELED: The solver thread has been canceled

This is based glib g_thread APIs so one must use a GMainLoop context in
order to use the asynchronous solver.