Add gui_cleanup() function called after gtk_main()
This adds a cleanup function that will unref the widgets used in
maze_solver_cb(). This avoids Gtk-CRITICAL messages on exiting while the
solver is running. This may occur if maze_solver_cb() is called after
the widgets have been destroyed but before gtk_main() returns. By taking
a ref on these widgets we ensure they will be valid in such case.