Add maze_get_neighbour_cell() function
This function will replace the ugly neighbours array in various
functions, starting with maze_create() in this patch.

The function takes a struct Cell and a Direction and returns a pointer
to the neighbour cell in the passed direction. It returns NULL if the
given direction goes outside the maze boundaries.

The function maze_get_neighbour_cell() returns an adjacent cell while
maze_get_neighbour_cell_offset() returns the cell at passed offset.
Calling maze_get_neighbour_cell() is equivalent to calling
maze_get_neighbour_cell_offset() with an offset of 1.